#118-refactor-SelectedUngredients-component
