--- front/src/components/SelectedRecipeList/SelectedRecipeList.module.scss
@@ -1,47 +1,0 @@
-.container {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px;
-  padding: 0;
-  max-width: 1033px;
-  margin: 20px 0 0 0;
-}
-
-.item {
-  background-color: #e9f9e2;
-  border-radius: 16px;
-  padding: 9px 14px 9px 20px;
-  display: flex;
-  gap: 12px;
-  font-size: 16px;
-  font-family: Mulish;
-  font-weight: 300;
-  line-height: 24px;
-  align-items: center;
-  height: fit-content;
-}
-
-.remove_btn {
-  display: flex;
-  gap: 6px;
-  align-items: center;
-  cursor: pointer;
-  height: 38px;
-
-  & p {
-    font-family: Mulish;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 300;
-    line-height: 20px;
-    color: #232323;
-    margin: 0;
-  }
-}
-
-.cross_btn {
-  width: 20px;
-  height: 20px;
-  background-color: transparent;
-  border: none;
-}
