+++ front/src/components/UI/Input/Input.module.scss
@@ -54,7 +54,8 @@
   }
   &_type_home {
     width: 100%;
-    background: url('images/header_search.svg') center right 16px no-repeat;
+    background: url('assets/images/header_search.svg') center right 16px
+      no-repeat;
     height: 46px;
   }
