+++ front/src/styles/_color-variables.scss
@@ -6,6 +6,7 @@ $color-secondary: #909090;
 $input-border-color: #e5e4e4;
 
 $chips-bg-color: #e9f9e2;
+$btn-chips-cross: #2c7429;
 
 // Переменные кнопок
 $btn-bg-primary: #f1b823;
